pyproject.toml
[project]
name = "llm-rcg24-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.10.0",
    "frozenlist>=1.5.0",
    "langchain>=0.2.5",
    "langchain-community>=0.2.5",
    "langchain-core>=0.2.8",
    "langchain-huggingface>=0.0.3",
    "multidict>=6.1.0",
    "pydantic>=1.10.21",
    "pydantic-core>=2.29.0",
    "pypdf>=5.3.0",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.1",
    "tiktoken>=0.8.0",
]
